refactor: refactoring to use project architecture
