+++ src/pages/Dashboard/Hotel/index.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import UserContext from '../../../contexts/UserContext';
 import useHotel from '../../../hooks/api/useHotel';
 import { Status, Title } from '../Payment/payment-card';
+import { ListHotels } from '../Hotel/listHotels';
 
 export default function Hotel() {
   const { hotels } = useHotel();
@@ -13,11 +14,6 @@ export default function Hotel() {
     hotelId: null,
   });
 
-  function handleHotel(hotelId) {
-    setData({
-      hotelId,
-    });
-  }
   //console.log(data);
 
   return (
@@ -28,22 +24,7 @@ export default function Hotel() {
 
       {hotels ? (
         <Container>
-          {hotels.map((hotel, key) => {
-            let single, double, triple = '';
-            hotel.Rooms.filter((cpt) => {
-              if (cpt.capacity === 1) single = 'Single,';
-              if (cpt.capacity === 2) double = 'Double,';
-              if (cpt.capacity === 3) triple = 'Triple';
-            });
-            return (
-              <StyledHotel key={key} onClick={() => handleHotel(hotel.id)}>
-                <Image src={hotel.image}></Image>
-                <NameHotel>{hotel.name}</NameHotel>
-                <SubTitle>Tipos de acomodação:</SubTitle>
-                <StyledSubTitle>{`${single} ${double} ${triple}`}</StyledSubTitle>
-              </StyledHotel>
-            );
-          })}
+          <ListHotels hotels={hotels} setData={setData} data={data} />
         </Container>
       ) : (
         ''
@@ -57,58 +38,4 @@ const Container = styled.div`
   width: 100%;
 
   display: flex;
-`;
-
-const StyledHotel = styled.div`
-  height: 264px;
-  width: 196px;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  padding: 15px;
-
-  margin-top: 20px;
-  margin-right: 20px;
-
-  background: #ebebeb;
-  border-radius: 10px;
-`;
-
-const Image = styled.img`
-  width: 168px;
-  height: 109px;
-
-  border-radius: 5px;
-`;
-
-const NameHotel = styled.div`
-  width: 100%;
-
-  font-family: 'Roboto';
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 23px;
-
-  color: #343434;
-
-  margin-top: 10px;
-`;
-
-const SubTitle = styled.div`
-  width: 100%;
-
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 14px;
-
-  color: #3c3c3c;
-
-  margin-top: 13px;
-`;
-
-const StyledSubTitle = styled(SubTitle)`
-  font-weight: 400;
-  margin-top: 5px;
 `;
